+++ src/windows/main_windows.cpp
@@ -358,7 +358,7 @@ void ts_platform_list_files_block(ts_search_result* result, wchar_t* start_dir)
 			char *matched_filter = 0;
 			utf8_int8_t uni_name[MAX_INPUT_LENGTH];
 			WideCharToMultiByte(CP_UTF8,0,name,-1,(LPSTR)uni_name,MAX_INPUT_LENGTH, NULL, NULL);
-			if (ts_filter_matches(&result->filters, uni_name, &matched_filter) == -1) {
+			if (ts_filter_matches(&result->filters, uni_name, &matched_filter) == (size_t)-1) {
 				continue;
 			}
 			(void)matched_filter;
